Use `point` inside mousehandler, instead of raw x and y values
Doesn't do much but improve continuity. I assume at some point
we might have a function that deals with points, and it would be
annoying if we had to reach inside mouse and create a point with
its values. It's easier if we just store them as a point from the
beginning.