Create `controllers' which handle key bindings
Explanation:

We need an abstraction on event handlers which allows us to easily
process keypresses, and automatically call functions based on which
key was pressed.

To do this we create a `binding', which is an object that contains
a reference to a key (string), and a function (state -> void).

We then create a `controller', which is a closure over an array of
bindings and an array of strings. The array of strings should be
created and updated async by an event handler.

`controller' contains a public reference to a function, `process',
which is (state -> void). This function iterates through the
bindings, checking to see if their corresponding keys are pressed.
If they are, it will call the function contained inside the
binding with the the state.

The result is that all we have to do to run all the right functions
in our update loop is
    controllers.forEach((controller) => {
        controller.process(state);
    });
All the other work (setting up keybindings, checking which key is
pressed, ...) is done elsewhere.

A nice side effect of this is that it allows us to more easily
remap keybindings during execution (which we may want to implement
some time in the future).