Refactor controller
I noticed that controller didn't work the way the rest of the
program does; by having one 'create' function that returns an
object, then other functions which take that object and other
arguments. Controller used to not do this, and instead bind a
function to the object itself before returning it from create.